support PartitionSplitter
